lol postgres

+++ db/migrate/20190316034359_add_confidence_to_trigger.rb
@@ -1,5 +1,5 @@
 class AddConfidenceToTrigger < ActiveRecord::Migration[5.0]
   def change
-    add_column :triggers, :confidence, :double
+    add_column :triggers, :confidence, :float
   end
 end
